#10 refactor modal
