+++ frontend/src/styles/components/ModalPaints.scss
